Implement very first version of seekable stream (#15)
Implement very first version of seekable stream

In this commit we are adding the first version of seekable stream that calls
the object client under the hood. For now, seeking is extremely trivial and
reading data amounts to proxying down read calls to the GetObject stream.

The testing part is more important. On top of unit tests we set up the
ability to test against a mocked version of S3. In the test scope we also
add an in-memory stream backed by a byte array. Reading out of this stream
and seeking in this stream is very easy to implement correctly and this can
provide a good base for ensuring correctness in the future.

We also add Checkstyle to enforce Javadoc comments.