Implement first version of JMH microbenchmarks (#18)
Now that we have a functional (correct but slow) implementation of
S3SeekableStream, it makes sense to put monitoring over performance.

This PR implements basic microbenchmarks that test full sequential read,
forward seeks, backward seeks and a Parquet-like ("jumping around")
pattern. For now, we only compare against the performance of a single
standard (=non-CRT) S3 async client. Using these benchmarks we can start
implementing optimisations and get relatively quick feedback of what (if
anything) they improved.

To run the microbenchmarks one has to assume AWS credentials and specify
two environment variables (a BUCKET and a PREFIX). We include a generator
utility so that setup is easy and this can later by open-sourced. The
README is updated with instructions about running these benchmarks.