Implement readTail() (#20)
This change works towards the S3FileIO - S3SeekableStream integration. This
should be the last change required to support S3FileIO.

readTail(buf, off, n) is a blocking method that reads all `n` bytes
requested before returning and does not alter the position in the seekable
stream. We had no equivalent for this so far, because if we wanted to
implement this with the existing S3SeekableStream APIs, then we'd need to
know the object size upfront -- an assumption we don't want to make
(probably shouldn't).

I see this method become useful later to implement tail pre-read: however,
a more efficient implementation would be similar to what S3FileIO is doing
today (pushing down the `bytes=-N` range into the S3 GET request). This
implementation is quick and dirty in that it relies on the content length
from the initial HeadObject we are making. We can spare this HeadObject in
the future once our integrations are running (and we have data that this is
worth doing).