Add property based testing (#34)
This change adds the ability to test properties rather than randomly
thought up test scenarios.

This comes with several benefits, a few of which are:

1\ Thinking up random test scenarios is error prone and tiring. Usually,
developers test their code by "example". In a better case, developers
parameterise their tests to raise confidence and increase coverage. Even
then, "tricky" corner cases are often forgotten. For example, programs
using the "double" floating point type will rarely be tested for values
such as -0.0, NaN or negative infinity. Property based testing frameworks
come with "corner case knowledge" and alleviate the need of needing to
manually specify these cases.

2\ Succintness. Goes in line with the above, but property based tests are
shorter and easier to maintain and read. If done right, they can be a useful
part of dev documentation and can provide a short specification in code.

3\ Learn as you go. On top of deterministic corner case behaviour, programs
are also tested using random input. This further decreases the chance of
forgetting useful test cases. The Java framework of choice (jqwik) is
capable of remembering violating corner cases (via a file-based database)
so that these are definitely tested in the future as well.

Finally, as a note: we do not propose that this is the holy grail of
testing. Our coverage should still be driven by unit tests (as it is the
case, this is why this is in a separate module). However, talking in terms
of properties should also help us introducing the right set of unit tests.