[Refactor] Move Parquet awareness out of the PhysicalIO layer (#56)
### Overview

- This is a pure refactor PR. It is not adding new functionality. It does
add new abstractions/new components though.

- This change moves Parquet-related code back into the LogicalIO layer. We
were planning for this change, Parquet awareness was only temporarily in
PhysicalIO after merging prototype code into mainline.

### Logical IO

- We introduce a new ParquetStatistics class that is meant to keep around
Parquet usage information and is meant to be shared among streams. It lives
next to MultiObjectsBlockManager in the Seekable Stream Factory, so the
factory is capable to inject it into new streams.

- We introduce a new component, called the ParquetPrefetcher. This makes it
possible to move Parquet detail out of the immediate vicinity of LogicalIO
and might help us move Parquet related stuff in its own module later. This
modularisation also allows us to easier test ParquetPrefetcher together
with LogicalIOConfiguration (as we do not need to mock all the way down, it
is enough to look at interactions with mock tasks).

- Exception throwing was revisited. An important aspect of the prefetching
is that it happens async and exceptions from it should not surface into the
main thread. However, asynchrony can also make it hard to test logic
enclosed by it, so we first make `*Tasks` synchronous and allow exceptions
to be thrown. We wrap checked exceptions into unchecked CompletionExceptions
(this works well with CompletableFutures, by default CompletableFutures
would wrap all exceptions into CompletionExceptions when you call join).
This in turn allows us to get rid of a lot of code that mimicked exceptional
behaviour via `Optional<..>`s and this now tests easier.

- All task results eventually end up in a CompletableFuture so should not
fail the main flow. Today, we do not call get or join on results.

### Physical IO

- Changes here are mostly deletes. There is no Parquet-related shared state
now here.

- Changed the void execute(IOPlan) "API" call on the PhysicalIO interface.
It returned void before, so there was no feedback mechanism available for
a caller to track the status of IOPlan-executions. Also, it helps testing
when you actually get back something from an async call and can join on
it/use the result. It can eliminate a lot of noise and flaky behaviour
because synchronising stuff becomes possible.

- Also, Ahmar was signalling for a while that the logical layer needs to be
aware how much is actually being read. I don't know if this can help in the
future, but it might.