Implement sequential prefetching in PhysicalIO (#65)
### Overview

The main contribution of this change is the addition of sequential
prefetching.  This seems to improve AQBE performance of S3FileIO by roughly
5%.  To achieve this, significant changes had to be made in the Physical IO
layer.

### Important PhysicalIO changes

One of the most important changes is that creating new IOBlocks is not sync
anymore.  This allows us to split huge ranges up into multiple smaller
blocks and fetch them all at the same time, thus benefitting from
parallelism.

We introduce more clever range managing to prevent double-reads. This lives
in its own component for the sake of easier testing.  We add logic to adjust
ranges once sequential read patterns are detected.  Within this, when a
range grows past a threshold we split it back to smaller chunks
(configurable) to benefit from parallelism.

### Architecture

We’ve made progress towards separating data caching and metadata caching by
implementing a MetadataStore and a BlobStore.

### Performance

Microbenchmarks now run in less than 5 minutes. Sequential reads are around
100x quicker on large objects when a 4KB buffer is used.

### Correctness

We do not have integration tests, so interesting regression tests were
coded up as microbenchmarks.  This way, they are not lost and can live
somewhere (better than nothing!) until we add integration tests.  They
provide a good way to easily reproduce issues.

Reference tests were hugely useful for verifying the new implementation.

### Observability

I was tuning/testing with collecting read traces and measuring IO stalls on
different read patterns in different queries.  The code for this is not
included in this PR.

### Misc
We make ranges always closed. There isn't even a way in code to specify an
open range anymore.

### Next steps
- Microbenchmarks we should start running as part of CI/CD
- Need to move correctness tests from microbenchmarks to integration tests
- Fine tuning
- Memory management is still naive. The scheduling is still naive.
  I haven’t seen these biting us too much yet but we should probably take a
  look at some point.

### Testing
- Microbenchmarks are working
- Reference tests are passing
- Did refactoring to improve our magic with ranges. Separated out that
  logic into components with responsibilities to be able to test better.
- Did both S3A and S3FileIO runs
- We have done an EMR + S3A run
- We are working on EMR + S3FileIO runs, did not happen before this PR
  got merged.