[Bugfix] Sequential prefetching should not shrink ranges (#126)
This commit ensures that we never choose a sequential prefetch range that
has a smaller length than the currently requested read. If the base of the
sequential prefetch progression (i.e., the base of the geometric
progression) is small enough, then it is possible that the range for
generation X is smaller than the currently requested read.

This can be easily solved by adding a `Math.max` to the right place and
adding a regression test.