[Pure refactor] Add `TelemetryFormat` (#154)
This commit does a couple of things and is a precursor to a JSON telemetry  
change.                                                                     
                                                                            
The main contribution is the addition of the TelemetryFormat interface.     
Before this change, there was no good mechanism to specify different ways    
to render telemetry. With this change, I am keeping the current, i.e.,      
human-readable version as default, no behaviour change is expected from     
this.                                                                       
                                                                            
Details for full picture:                                                   
- We add a default telemetry format which is essentially all the code that  
was responsible for formatting before but split out into its own class to   
separate responsibilities.                                                  
- We add a new TelemetryFormat interface                                    
- OperationMeasurement and MetricMeasurement DO NOT have a default          
`toString()` method anymore, you MUST specify a TelemetryFormatter for      
rendering -- this makes sense because now they are containers of data more  
than anything else and this data can be rendered in a bunch of different    
ways.