Remove mentions of old repository name (#168)
## Description of change

We changed the name of our repository and now need to fix up some
mentions of the repo in CICD files.

#### Relevant issues
N/A

#### Does this contribution introduce any breaking changes to the
existing APIs or behaviors?

Yes, but in a good direction. Right now, CICD is failing. It should pass
now.

#### Does this contribution introduce any new public APIs or behaviors?

No

#### How was the contribution tested?

Cannot run CICD locally, so just hoping this will move the needle.

#### Does this contribution need a changelog entry?
- [x] I have updated the CHANGELOG or README if appropriate

---

> By submitting this pull request, I confirm that my contribution is
made under the terms of the Apache 2.0 license and I agree to the terms
of the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).