Updated README.md (#171)
Minor readme edits for consistency.

## Description of change
Minor README edits

#### Relevant issues
None

#### Does this contribution introduce any breaking changes to the
existing APIs or behaviors?
No

#### Does this contribution introduce any new public APIs or behaviors?
No

#### How was the contribution tested?
Preview tab on GitHub

#### Does this contribution need a changelog entry?
Please advise if this is needed. 

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).