Add JMH JAR generation to CICD (#207)
This commit makes sure that we always generate the JMH JAR that is used
for running the microbenchmarks. Previously, this was an unexercised
workflow, which meant that we could break it without detecting it.

Including in CICD is essentially ensuring that we have our loops closed
on this and should never break microbenchmark builds.