feat: adding etag checking for stream reads (#209)
## Description of change
We want etags on our get requests to verify that object reads are for
the same version of the object as it is a durability concern. To do this
I'm adding etags to the metadata for an object and then making
conditional get requests.

Additionally pinning an objects metadata to a point in time. To stop us
getting the metadata at different time points during a read. This idea
helps us keep metadata consistent for a whole stream.

Additionally now am clearing the metadata for an object on exception.
This will change when I pull in
https://github.com/awslabs/analytics-accelerator-s3/pull/210. But just
to demonstrate the idea I made it on all exceptions at first.


#### Does this contribution introduce any breaking changes to the
existing APIs or behaviors?
Yes

#### Does this contribution introduce any new public APIs or behaviors?
Yes

#### How was the contribution tested?
* Unit tests:
For the unit tests: checking if we change the etag between blobs we get
an error and used argument matcher for this
for this file
[BlockTest.java](https://github.com/awslabs/analytics-accelerator-s3/compare/stubz151_etag?expand=1#diff-75cd3f62ba96f2b7d7e806e6b9c5572e63cbfb2e8eb1052ae02ca74ffa76cfb5)
I just updated the values to include the etag don't think it's that
important

* Integration tests:
For integration tests: added 2 one that verifies we fail when it changes
mid read and a second that validates we pass and return the cached
object if it's done reading.



#### Does this contribution need a changelog entry?
- [ ] I have updated the CHANGELOG or README if appropriate
TODO still, will update

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).