add gray failure tests and FaultyS3Client (#214)
## Description of change
This change adds a new S3Client to TestFixtures -- FaultyS3Client. 
The purpose of this client to inject failures to S3 Interactions and
test for resiliency of the library.

#### Relevant issues
AWS Java SDK V2 has an issue of getting stuck time to time:
https://github.com/aws/aws-sdk-java-v2/issues/5755
With this PR we are reproducing a similar behaviour. In the follow-up
PRs we will implement defense in depth mechanisms.

#### Does this contribution introduce any breaking changes to the
existing APIs or behaviors?
No.

#### Does this contribution introduce any new public APIs or behaviors?
No.

#### How was the contribution tested?
This is a test contribution itself. New tests are disabled.
Confirmed `./gradlew integrationTest` are still passing with the 2
trusted SDK clients and not triggered with the faulty client.


#### Does this contribution need a changelog entry?
No. 

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).