Reduce long to int type cast (#220)
## Description of change
When copying bytes from block content to buf passed in parameter, we
re-compute content offset for each byte.
This means doing a long -> int cast for every byte read from the stream.
This conversation is not needed to be done multiple times as start of a
block is final and pos is not changing inside the method. This PR
removes unnecessary conversation.

#### Relevant issues
<!-- Please add issue numbers. -->
<!-- Please also link them to this PR. -->

#### Does this contribution introduce any breaking changes to the
existing APIs or behaviors?
<!-- Please explain why this was necessary. -->

#### Does this contribution introduce any new public APIs or behaviors?
<!-- Please describe them and explain what scenarios they target.  -->

#### How was the contribution tested?
<!-- Please describe how this contribution was tested. -->

#### Does this contribution need a changelog entry?
- [ ] I have updated the CHANGELOG or README if appropriate

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).