Added extra logging to InputStream reader (#230)
## Description of change
This PR adds extra logging to StreamUtils to be able to debug the
timeout issues.

#### Relevant issues
- https://github.com/awslabs/analytics-accelerator-s3/pull/229
- https://github.com/awslabs/analytics-accelerator-s3/pull/219

#### Does this contribution introduce any breaking changes to the
existing APIs or behaviors?
No

#### Does this contribution introduce any new public APIs or behaviors?
No

#### How was the contribution tested?
- Unit tests

#### Does this contribution need a changelog entry?
No

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

---------

Co-authored-by: Erdogan Ozkoca <ozkoca@amazon.com>