add javadoc generation (#237)
## Description of change
Fixing some issues with javadoc generation and restricting javadoc to
only supported interface.

#### Relevant issues
N/A

#### Does this contribution introduce any breaking changes to the
existing APIs or behaviors?
No

#### Does this contribution introduce any new public APIs or behaviors?
No

#### How was the contribution tested?
./gradlew clean build
./gradlew javadoc
./gradlew integrationTests

#### Does this contribution need a changelog entry?
- Yes.

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).