fix: Ensure proper project evaluation for javadoc task (#241)
## Description of change
This makes sure build does not fail even when the signing is enabled.

#### Relevant issues
NA

#### Does this contribution introduce any breaking changes to the
existing APIs or behaviors?
No

#### Does this contribution introduce any new public APIs or behaviors?
No

#### How was the contribution tested?
`./gradlew clean build`

#### Does this contribution need a changelog entry?
No

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).