Add debug log for S3SeekableInputStreamConfiguration (#250)
## Description of change
This adds the debug log of the configuration when the
S3SeekableInputStreamFactory is initialised. It helps us to verify if
the config passed by the connectors is rightly consumed.

#### Relevant issues
NA

#### Does this contribution introduce any breaking changes to the
existing APIs or behaviors?
No

#### Does this contribution introduce any new public APIs or behaviors?
No

#### How was the contribution tested?
Ran an existing integration test to see its being logged properly.

#### Does this contribution need a changelog entry?
No
---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).