improve integration testing documentation (#260)
## Description of change
Minor changes to improve documentation for Integration Testing. 

#### Relevant issues
N/A

#### Does this contribution introduce any breaking changes to the
existing APIs or behaviors?
No, this is only changing documentation

#### Does this contribution introduce any new public APIs or behaviors?
No, this is only changing documentation

#### How was the contribution tested?
N/A

#### Does this contribution need a changelog entry?
- [X] I have updated the CHANGELOG or README if appropriate

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).