Migration to Central Portal for Maven (#282)
## Description of change
This PR makes the relevant changes to migrate maven publish to the
[central portal](https://central.sonatype.com/account). Since we are
depending on the `maven-publish` plugin, we also need to update our
release-workflow to make a POST request to use UI for the next steps in
release as mentioned
[here](https://central.sonatype.org/publish/publish-portal-ossrh-staging-api/#post-to-manualuploaddefaultrepositorynamespace).

#### Relevant issues
NA

#### Does this contribution introduce any breaking changes to the
existing APIs or behaviors?
No

#### Does this contribution introduce any new public APIs or behaviors?
No

#### How was the contribution tested?
Manual Release 

#### Does this contribution need a changelog entry?
No

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).