Upgrade to freefair lombok version 8.14 (#313)
## Description of change

upgrade to latest version of FreeFair Lombok: 8.14

https://github.com/freefair/gradle-plugins/releases


#### Relevant issues

n/a

#### Does this contribution introduce any breaking changes to the
existing APIs or behaviors?

n/a

#### Does this contribution introduce any new public APIs or behaviors?

n/a

#### How was the contribution tested?

CI test suite

#### Does this contribution need a changelog entry?

no

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

[ x] yes