Upgrade SDK version and fix user-agent changes to avoid prepending empty string (#319)
This PR does 2 things 
1/It updates SDK version to the same version as Iceberg. This is because
SDK introduced a breaking change on user-agent evaluation.
2/It adds a small check to User agent, does not prepend if string is
empty. It confirms client can still make requests with different
ObjectClientConfigurations (minimal checks to be fair) to ensure we can
still read objects with different configurations.
3/ It adds CRCCheck class, taken out from SDK to confirm streams read
from S3 are identical with and without AAL.



## Description of change
<!-- Thank you for submitting a pull request!-->
<!-- Please describe your contribution here. What and why? -->
<!-- Please ensure your commit messages follow these
[guidelines](https://chris.beams.io/posts/git-commit/). -->

#### Relevant issues
<!-- Please add issue numbers. -->
<!-- Please also link them to this PR. -->

#### Does this contribution introduce any breaking changes to the
existing APIs or behaviors?
<!-- Please explain why this was necessary. -->

#### Does this contribution introduce any new public APIs or behaviors?
<!-- Please describe them and explain what scenarios they target.  -->

#### How was the contribution tested?
<!-- Please describe how this contribution was tested. -->

#### Does this contribution need a changelog entry?
- [ ] I have updated the CHANGELOG or README if appropriate

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).