Upgrade to junit 5.13.4 (#314)
## Description of change

upgrade junit to version 5.13.x

## Motivation

junit 5.13.x has a new annotation:  ParameterizedClass


https://docs.junit.org/current/api/org.junit.jupiter.params/org/junit/jupiter/params/ParameterizedClass.html

#### Relevant issues

n/a

#### Does this contribution introduce any breaking changes to the
existing APIs or behaviors?

n/a

#### Does this contribution introduce any new public APIs or behaviors?

n/a

#### How was the contribution tested?

existing test suite

#### Does this contribution need a changelog entry?
- [ ] I have updated the CHANGELOG or README if appropriate

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

[x] yes

---------

Co-authored-by: Fuat Basik <fuatbasik@gmail.com>